perf(users): optimize user listing with single query and projections

+++ backend/src/main/java/ua/lviv/bas/cinema/domain/projection/AdminUserProjection.java
@@ -26,7 +26,7 @@ public interface AdminUserProjection {
 
 	LocalDateTime getUpdatedAt();
 
-	int getTicketsCount();
+	Long getTicketsCount();
 
 	LocalDateTime getLastActivity();
 }
